sep第一个版本

+++ huaji/index.js
@@ -4,27 +4,58 @@ const shelljs = require('shelljs')
 const chalk = require('chalk')
 var babel = require('@babel/core')
 
+const exists = fs.existsSync
+
 function FangTianHuaJi() {
   this.cssJs = `import './index.css'\nimport '../../style/index.css'`
 }
 
 FangTianHuaJi.prototype.excute = function() {
-  glob.sync('src/components/*/style/*.tsx').forEach(path => {
+  if (!exists('sep/es')) {
+    shelljs.mkdir('-p', 'sep/es')
+    shelljs.cp('src/components/index.js', 'sep/es/index.js')
+  }
+  glob.sync('src/components/*/style/{*.tsx,*.less}').forEach(path => {
+    const isLess = /.less$/.test(path)
     const cssOutputDir = `${path
-      .replace(/^src/, 'dist')
+      .replace(/^src/, 'sep')
       .replace(/components/, 'es')
       .split('/')
       .splice(0, 4)
       .join('/')}`
-
     const cssOutputPath = `${cssOutputDir}/css.js`
 
-    shelljs.mkdir('-p', cssOutputDir)
-    shelljs.touch(`${cssOutputPath}`)
+    if (!exists(cssOutputDir)) {
+      shelljs.mkdir('-p', cssOutputDir)
+    }
+    if (!exists(cssOutputPath)) {
+      shelljs.touch(`${cssOutputPath}`)
+      fs.writeFileSync(cssOutputPath, this.cssJs, err => {
+        err && console.log(chalk.red(err))
+      })
+    }
 
-    fs.writeFileSync(cssOutputPath, this.cssJs, err => {
-      err && console.log(chalk.red(err))
-    })
+    if (isLess) {
+      const lessOutputPath = path
+        .replace(/^src/, 'sep')
+        .replace(/components/, 'es')
+      shelljs.cp(path, lessOutputPath)
+    }
+  })
+
+  glob.sync('src/components/style/**/*.less').forEach(path => {
+    const lessOutputPath = path
+      .replace(/^src/, 'sep')
+      .replace(/components/, 'es')
+    const len = lessOutputPath.split('/').length
+    const lessOutputDir = lessOutputPath
+      .split('/')
+      .splice(0, len - 1)
+      .join('/')
+    if (!exists(lessOutputDir)) {
+      shelljs.mkdir('-p', lessOutputDir)
+    }
+    shelljs.cp(path, lessOutputPath)
   })
 }
 
@@ -37,7 +68,7 @@ FangTianHuaJi.prototype.tail = function() {
 
     fs.readFile(path, (err, data) => {
       const outputPath = path
-        .replace(/^src/, 'dist')
+        .replace(/^src/, 'sep')
         .replace(/components/, 'es')
         .replace(/.tsx$/, '.js')
       err && console.log(chalk.red(err))
@@ -47,17 +78,19 @@ FangTianHuaJi.prototype.tail = function() {
     // lib version
     babel.transformFile(path, {}, function(err, result) {
       const outputPath = path
-        .replace(/^src/, 'dist')
+        .replace(/^src/, 'sep')
         .replace(/components/, 'lib')
         .replace(/.tsx$/, '.js')
       err && console.log(chalk.red(err))
       fs.writeFileSync(outputPath, result.code)
     })
   })
+  shelljs.cp('./package.json', './sep/package.json')
+  shelljs.cp('./package-lock.json', './sep/package-lock.json')
 }
 
 FangTianHuaJi.prototype.clear = function() {
-  shelljs.rm('-rf', 'dist')
+  shelljs.rm('-rf', 'sep')
 }
 
 const fangTianHuaJi = new FangTianHuaJi()
